test(j1939): drop a redundant cast CodeQL flagged on #139

+++ tests/CanKit.Pro.Tests/TestCases/J1939/J1939NodeTests.cs
@@ -338,7 +338,7 @@ public class J1939NodeTests : IClassFixture<VirtualAdapterFixture>
         // Unclaimed: Cannot-Claim.
         var seen = WaitForMessageAsync(node, m => J1939Pgn.IsRequest(m.Pgn), ShortTimeout);
         Request();
-        (await NextClaimAsync()).Should().Be(((byte)J1939Pgn.NullAddress, nodeName.Value), "a node without an address answers Cannot-Claim");
+        (await NextClaimAsync()).Should().Be((J1939Pgn.NullAddress, nodeName.Value), "a node without an address answers Cannot-Claim");
         (await seen).SourceAddress.Should().Be((byte)0x20, "the request still reaches the application");
 
         // Claimed: the claim itself.
